JPG: Added jpgStructs file
