+++ src/codecFunctions/jpg/jpgManagement/jpgComponentManagement.hpp
@@ -1,12 +1,4 @@
-
-
-struct Component{
-    unsigned char componentID;
-    unsigned char sampligFactors;
-    unsigned char quatizationTable;
-    unsigned char huffmanTableDC;
-    unsigned char huffmanTableAC;
-};
+#include "jpgStructs.hpp"
 
 
 void assignHuffTablesToComponent(unsigned char huffTableByte, Component & component);
